+++ src/main/java/com/example/oneulnail/domain/user/dto/response/SignMessageResDto.java
@@ -1,4 +1,4 @@
-package com.example.oneulnail.domain.user.dto.sign;
+package com.example.oneulnail.domain.user.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,9 +8,6 @@ import lombok.ToString;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@ToString
-public class SignUpResDto {
-
-    private String msg;
-
+public class SignMessageResDto {
+    private String authenticationNumber;
 }
